checklist updated AnswerDTO, AnswerConverter (added new field)
analytics
created new event rules
added new constants
added builders functional to QueryRequest
created dao functional CypherQueryService
updated dtos
created new exception CypherOperationException
added new functional for mining fact questions

+++ analytic-app/src/main/java/com/datapath/analyticapp/dto/imported/response/ConditionCharacteristicDTO.java
@@ -5,5 +5,5 @@ import lombok.Data;
 @Data
 public class ConditionCharacteristicDTO {
     private Long outerRiskEventId;
-    private Integer conditionAnswerId;
+    private Long conditionAnswerId;
 }
